\name{geopath}
\alias{geopath}
\title{Vega geopath}
\usage{
  geopath(fill.color = "pink", hover.color = "lightblue",
    border.color = "black", url = "data/us-states.json",
    projection = "albersUsa", scale = 500,
    translate = c(a$opt$width/2, a$opt$height/2), ...)
}
\description{
  Create a geopath visualization
}

