\name{renderVega}
\alias{renderVega}
\title{renderVega (use with Shiny)}
\usage{
  renderVega(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
  \item{expr}{An expression that returns a chart object}

  \item{env}{The environment in which to evaluate
  \code{expr}.}

  \item{quoted}{Is expr a quoted expression (with
  \code{quote()})? This is useful if you want to save an
  expression in a variable.}
}
\description{
  ...
}

