\name{Vega}
\alias{Vega}
\title{Vega class}
\description{
  ...
}
\section{\code{html()}}{
  A method to convert the chart object to HTML code. The
  code does not include necessary JavaScript files (e.g.
  Vega and jQuery); however, when used interactively
  (\code{show()}) or with Shiny, those are automatically
  included.
}
\examples{
\dontrun{
...
}
}

