\name{vega_geopath}
\alias{vega_geopath}
\title{Vega geopath}
\usage{
  vega_geopath(data, labels = names(data), width = 1200,
    height = 800, font = "Helvetica Neue", font_size = 14,
    fill_color = "pink", hover_color = "lightblue",
    padding = c(0, 0, 0, 0), url = "data/us-states.json",
    projection = "albersUsa", scale = 300)
}
\description{
  Create a geopath visualization
}

