\name{geopath}
\alias{geopath}
\title{Vega geopath}
\usage{
  geopath(data, labels = names(data), width = 1200,
    height = 800, font = "Helvetica Neue", font.size = 14,
    fill.color = "pink", hover.color = "lightblue",
    border.color = "black", padding = c(0, 0, 0, 0),
    url = "data/us-states.json", projection = "albersUsa",
    scale = 300, translate = c(width/2, height/2))
}
\description{
  Create a geopath visualization
}

