\name{geopath}
\alias{geopath}
\title{Vega geopath}
\usage{
  geopath(data = NULL, id = names(data), stroke = "white",
    strokeWidth = 0.5, hover.stroke = stroke,
    hover.strokeWidth = strokeWidth,
    url = "json/wdi.geo.json", projection = "orthographic",
    scale = 800,
    translate = c(a$opt$width/2, a$opt$height/2), ...)
}
\description{
  Create a geopath visualization
}

