\name{vega_geopath}
\alias{vega_geopath}
\title{Vega geopath}
\usage{
  vega_geopath(data, labels = names(data), width = 500,
    height = 400, font = "Helvetica Neue", font_size = 14,
    fill_color = "pink", hover_color = "lightblue",
    padding = c(0, 0, 0, 0),
    url = "data/world-countries.json",
    projection = "winkel3")
}
\description{
  Create a geopath visualization
}

